major overhaul of values: 1- from btc double to satoshi int64; 2- rework merging tx graphs to block graph

+++ BC2G/Blockchains/Bitcoin/Graph/C2SEdge.cs
@@ -16,13 +16,13 @@ public class C2SEdge : S2SEdge
     public new EdgeLabel Label { get; } = EdgeLabel.C2SMinting;
 
     public C2SEdge(
-        ScriptNode target, double value, uint timestamp, long blockHeight) :
+        ScriptNode target, long value, uint timestamp, long blockHeight) :
         base(
             ScriptNode.GetCoinbaseNode(), target,
             value, EdgeType.Mints, timestamp, blockHeight)
     { }
 
-    public new C2SEdge Update(double value)
+    public new C2SEdge Update(long value)
     {
         return new C2SEdge(Target, Value + value, Timestamp, BlockHeight);
     }
